+++ contract/src/tests/workspace/utils.rs
@@ -1,38 +1,34 @@
-use crate::Role;
 use near_sdk::serde_json::json;
 use near_workspaces::network::Sandbox;
 use near_workspaces::types::NearToken;
 use near_workspaces::{Account, AccountId, Contract, Worker};
 
+use crate::Role;
+
 const FACTORY_OWNER: &str = "factory-owner";
-const AURORA_FACTORY_CONTRACT_PATH: &str =
-    "target/wasm32-unknown-unknown/release/aurora_controller_factory.wasm";
-const INITIAL_BALANCE: NearToken = NearToken::from_near(200);
+const AURORA_FACTORY_CONTRACT_PATH: &str = "../res/aurora-controller-factory.wasm";
+pub const INITIAL_BALANCE: NearToken = NearToken::from_near(200);
 
 pub async fn crate_factory() -> anyhow::Result<(Account, Contract, Worker<Sandbox>)> {
     let worker = near_workspaces::sandbox().await?;
-    let root = worker.root_account().unwrap();
+    let root = worker.root_account()?;
     let factory_owner = root
         .create_subaccount(FACTORY_OWNER)
         .initial_balance(INITIAL_BALANCE)
         .transact()
-        .await
-        .unwrap()
-        .into_result()
-        .unwrap();
+        .await?
+        .into_result()?;
 
     let wasm = std::fs::read(AURORA_FACTORY_CONTRACT_PATH)?;
     let contract = factory_owner.deploy(&wasm).await?.result;
 
     let result = factory_owner
         .call(factory_owner.id(), "new")
-        .args_json(json!({"owner_id": factory_owner.id()}))
+        .args_json(json!({"dao": factory_owner.id()}))
         .transact()
-        .await
-        .unwrap();
-    assert!(result.is_success());
+        .await?;
+    assert!(result.is_success(), "{result:#?}");
 
-    grant_role(&contract, &factory_owner, factory_owner.id(), Role::DAO).await;
     grant_role(&contract, &factory_owner, factory_owner.id(), Role::Updater).await;
 
     Ok((factory_owner, contract, worker))
@@ -53,5 +49,5 @@ pub async fn grant_role(
         .transact()
         .await
         .unwrap();
-    assert!(result.is_success());
+    assert!(result.is_success(), "{result:#?}");
 }