+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 2024-12-02
+
+- Reworked the `get_deployments` view method.
+
+## 0.2.0 2024-12-02
+
+- Update dependencies.
+- Make transaction payable.
+- Do not update deployment info after failed upgrade.
+- Remove ownable functionality.
+- Add a possibility to set amount of gas for migration.
+- Use enum for collection prefixes.
+
 ## 0.1.0 2023-12-22
 
 - Initial implementation. See the `README.md` for more details.